backup/requirements.txt
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pymongo==4.6.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
certifi==2024.2.2